+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 -   Expose a JSON-RPC client, using the wallet's current JSON-RPC server.
+-   `getSelectedAccount` method, which allows dApps to get the current account with using `connect`.
 
 ### (Breaking) Changed